Revert "LPS-119068 Create intermediate component to setup listeners al SideNavigation"
This reverts commit 124dd8a55db74fa4647d95fcedb31f33a23111cb.
